+++ cmd/web/api/user.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	"github.com/ostamand/url/cmd/web/helper"
-	"github.com/ostamand/url/cmd/web/store"
+	"github.com/ostamand/url/internal/store"
 )
 
 type UserController struct {
