+++ cmd/web/api/user_test.go
@@ -8,10 +8,10 @@ import (
 	"os"
 	"testing"
 
-	"github.com/ostamand/url/cmd/web/config"
 	"github.com/ostamand/url/cmd/web/helper"
-	"github.com/ostamand/url/cmd/web/store"
-	"github.com/ostamand/url/cmd/web/store/mysql"
+	"github.com/ostamand/url/internal/config"
+	"github.com/ostamand/url/internal/store"
+	"github.com/ostamand/url/internal/store/mysql"
 )
 
 var ctrl *UserController
