+++ cmd/web/handler.go
@@ -8,7 +8,7 @@ import (
 
 	ctrl "github.com/ostamand/url/cmd/web/controller"
 	"github.com/ostamand/url/cmd/web/helper"
-	"github.com/ostamand/url/cmd/web/store"
+	"github.com/ostamand/url/internal/store"
 )
 
 type Handler struct {
