+++ cmd/web/helper/user_helper_test.go
@@ -7,9 +7,9 @@ import (
 	"testing"
 	"time"
 
-	"github.com/ostamand/url/cmd/web/config"
-	"github.com/ostamand/url/cmd/web/store"
-	"github.com/ostamand/url/cmd/web/store/mysql"
+	"github.com/ostamand/url/internal/config"
+	"github.com/ostamand/url/internal/store"
+	"github.com/ostamand/url/internal/store/mysql"
 	"github.com/stretchr/testify/assert"
 )
