+++ internal/store/mysql/mysql_test.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	"github.com/DATA-DOG/go-sqlmock"
-	"github.com/ostamand/url/cmd/web/store"
+	"github.com/ostamand/url/internal/store"
 	"github.com/stretchr/testify/assert"
 )
